BUG: Move legacy check for void printing
The check needs to be in the python path, because also the printing
for `str()` subtly changed.

To be the same as tuples, we should actually print `(np.float32(3.), np.int8(1))`
for `str()` (tuples include the repr), while for `repr()` we include the dtype
so we would print similar (but ideally with full precision) to arrays.

This isn't quite ideal, I would be happy to print the full repr in the `str()`
but I guess that might be a bit annoying in practice, so maybe our Numeric types
are special enough for now.