API: Remove The MapIter API from public
This API is very slow to use compared to normal advanced indexing
(which reaches deep into the much richer internals).
Making this fast is difficult and the only known users is Theano:
* While probably still in use Theano is end-of-life for a long time.
* The use-case in Theano would be better served with `ufunc.at` or
  direct advanced indexing use probably (even if that requires
  wrapping into NumPy arrays).  Because it will be vastly faster
  normally.

So remove it (make it private) to allow making the ABI of it not be
a strange mix for backwards compatibility.