ENH: Use a pointer buffer to handle advanced integer indexing
This calculates pointer offsets (in most paths) of advanced indexing
rather than integrating the calculation into the code where we use
the index.
Thus, it simplifies the code, although I retain it for the moment for
the simplest case of advanced indexing (mapiter_trivial_get/set) because
it may be a larger speed difference there.