VENDOR: vendor the Meson build system (with modifications for numpy)
This vendors the current state of https://github.com/mesonbuild/meson/pull/11307,
which is the "feature module" for the multi-target build support that
NumPy needs to build its SIMD support (by Sayed Adel). That PR is,
as vendored, based on top of the state of the Meson master branch
in commit 9d88d0d5c, which is a dev commit in between the 1.2.0 and
1.2.1 tags from 18 July 2023.

No modifications have been made to any of the vendored files.