MAINT: revert adding `distutils` and `array_api` to `np.__all__`
This has turned out to be too disruptive; `from numpy import *`
has to remain warning-free and these two modules emit a warning
on import rather than when they're actually used.

`array_api` we have to decide on for 2.0 (either remove the experimental
warning or remove it), I'll open a separate issue to follow up on that.