MAINT: test-refactor of `numpy/_core/numeric.py`
This is a pure-test refactor (as well as adding a few cases), extracted
from numpy/numpy#24663.

Testing each individually has a small Pytest overhead, but has the
advantage of each test being able to fail independently in case of a
modification of the tested function.

It also make it much easier to add a new test, and I believe is much more
readable.

[skip cirrus] [skip circle] [skip azp]