BUG: Fix weak promotion with some mixed float/int dtypes
This reorganizes/simplifies promotion a bit, but unfortunately this
means moving the promotion of our concrete DTypes with the abstract
"weak" ones into the concrete DTypes making the order of "who deals
with whom" fully defined by:

    complex' > floats > pycomplex > pyfloat > ints > pyint > bool

where py<...> is the DType corresponding to the Python type.
There are two points here:
* "inexact" is the meaningful category not complex
* previously, the abstract DTypes `pyfloat`, etc. knew about the
  concrete ones and there was some special handling for that.
  (as mentioned above, this is gone now)

Why was it this complicated?!  I think the initial code wanted to
deal with value-based promotion and the abstract DTypes could have
had a value attached to them.
There was thus more code later on, dealing with abstract DTypes
more concretely (rather than identical to the rest).

This is now gone, concrete DTypes promote basically the same as
their concrete counter-parts.  This makes the promotion logic simpler
(and I think still just as correct), although the concrete type
promotion now needs to know about it.