DOC: add a note on the ``.c.src`` format to the distutils migration page
Detailed migration guidance should probably be out of scope here,
since this is a rarely used feature. Not using the format at all
would be best, since it's a very poor format and we'd like to get of
it in NumPy itself. Better options include using C++ templates,
using a more standard templating library, or using C11's ``_Generic``.

[skip azp] [skip actions] [skip cirrus]