BUG, SIMD: use scalar cmul on bad Apple clang x86_64
Apple clang 14.0.0 outputs code with non-deterministic behaviour for the
AVX2-accelerated `multiply` ufunc on `complex64` and `complex128` for
x86_64 on macOS with AVX2 enabled.  This bug is fixed by Apple clang
14.0.3, but 14.0.0 is still commonly the available toolchain on CI
images.

In order to not output unsound code, this simply skips the SIMD version
of the ufunc when using an affected compiler.