DOC: how-to-io.rst: document solution for NumPy JSON serialization
Users have reported that NumPy types are not JSON serializable. This is documented, but no solution is recommended. This PR suggests trying a custom NumPy JSON encoder, several of which are easily found online. Whether NumPy should provide a solution as part of the package remains an open feature request.

[skip actions] [skip azp] [skip cirrus]