BLD: Skip MSVC compatibility when building against mingw-w64 Python
In MSYS2 we build numpy like Linux distros, so we prefer dynamic
linking and it needs to be ABI compatible with all other packages.
numpy by default statically links ucrt and changes the ABI to match
MSVC, which makes it incompatible with other packages.

To differentiate the two use cases detect if we are building against
a mingww-64 Python build, and if we are skip all the special adjustments.

Note that mingw-w64 is not officially supported by upstream CPython.