MAINT: Add `npy_2_compat.h` which is designed two work also if vendored
This adds a header to define compatibility with 1.x when compiling
against 2.0.  However, it is also designed so that it can be vendored
by downstream to allow affected projects to compile with 1.x also.

For now, it only points to the complex macros and includes the
``NPY_DEFAULT_INT`` macro.  More things could be added though,
for example, I had an initial version that "backported"
``PyUFunc_GiveFloatingPointErrors``.