BLD: change default of the `allow-noblas` option to true.
Having it set to false caused more disruption and complaints than
the silent performance regressions for `true` ever did.
So now that the dust has settled a bit on the transition to Meson,
go back to the more permissive default.

The warning that is emitted when BLAS and/or LAPACK are not found
is already much more visible than it was with distutils, so this
new situation is still an improvement over where we were until
1.25.x.

Closes gh-24703.