BUG: Avoid intp conversion regression in Cython 3
In the Cython 3 `__init__.pxd` `intp` was defined to `Py_ssize_t`
which is *correct*.  However, Cython has special rules for `Py_ssize_t`.
It uses the `PyNumber_AsIndex` (`operator.index`) function for conversion
which rejects for example floats.

Of course I prefer that behavior, but:
1. There is no great reason why `np.intp_t` should be all that special.
2. Changing it silently changes downstream (and `np.random`) behavior.

It would be awesome if Cython would allow transitioning to using
`operator.index()` style coercion since you can still explicitly
convert via `int()`.

Now, why the new definition to `signed long`!?  `ssize_t` doesn't work
because it is *also* special, using `PyLong_AsSize_t`.
It seems to me that adding signed might give this type a bit of
preference in promotion, so it is is just for good sport.
(It is impossible to get Cython promotion fully right.)