DEP: Deprecate `__array_prepare__`
We don't call array-prepare consistently only for normal ufuncs,
it doesn't seem used by many people (the main use-case seems units
and masked-arrays, both of which are clearly much better served
by `__array_ufunc__`, even a bad version of it).

Now, there *is* one dance/change that needs to be OK, if the user
does NOT implement `__array_prepare__`, we make it a true no-op
now, while previously it would have done `arr.view(type=subclass)`.

That matters in one case in the tests where a faulty `__array_wrap__`
implementation expected to get a subclass.