MAINT: Cleanup mapiter struct a bit
This is a small followup to making MapIter private.  It renames
`numiters` to `num_fancy` (because numiters only made sense man years ago).

Also reorders some fields (tried to move some less used ones to the end
and clump them a bit clearer) and rewrites basically all the docs to be
maybe half understandable.
(This stuff is confusing/complicated, but I think it is a lot clearer now
probably)