API: Shrink MultiIterObject and make `NPY_MAXARGS` a runtime macro
Cython is OK with increased sizes, but doesn't like shrinking an
object size.  We should fix this (providing a size to compare with
in the `pyd` maybe).  But just not including the end of the object
here publically is totally fine...

This makes `NPY_MAXARGS` also a *runtime* constant, since the
presumably only user (`numexpr`) should be better off with that
anyway.  Yes, they need to change the code to hard-code the maximum,
but they also need the NumPy cap from us anyway.

This is needed, because Cython breaks for sklearn which uses
`numpy.broadcast` AKA `PyArrayMultiIterObject`.