API: Remove `PyDataType_ISPYTHON` and related macros
We never use them, they are an odd concept, as they pretend our
doubles are Python doubles which they are not really (maybe once
upon a time).

The concept thus seems useless to me, nothing maps to a Python object
clearly, and all integers map to Python ints with `.item()`, so what
does it mean?