API: Remove `PyArray_FieldNames` as it has no known reason to exist
It probably is also wrong if you use it on a field dict directly
since it mangles up the order.  The logic was designed for a *new*
dtype creation from only a dict, and even there it is unclear that
we should use it, since Python guarantees dict order...

In either case, I can't imagine a reasonable use-case, we don't use it.