Performance improvement of polyutils.as_series
This small patch provides a (small) performance improvement: For the
normal (straight / no error) case the improvement is from 0 - 5%, for
the error case the improvement is between 3 and 30%. The improvement
highly depends on the value of the parameter.

The original code always runs through all arrays, even if it could
stop at the first found with size zero.

Signed-off-by: Andreas Florath <andreas@florath.net>