Run f2py's meson backend with the same python that ran f2py
On Debian, we sometimes have multiple Python 3 versions in the archive.
When using f2py3.X, I'd expect meson to run with the same python, not
the default python.

Not sure how far up the stack this should come from, I kept it simple
for this PR, if someone has better ideas, they can direct me.