DOC: devguide cleanup: remove Gitwash and too verbose Git details
Also clean up some outdated info on the CI setup and wheel builds.
This is a very straightforward PR, and already pre-discussed in the
community meeting 2 weeks ago. This change does not yet fix the more
structural issues with this developer guide, like it being quite hard to
navigate. Next step is to update all the doc build info, and after
that the devguide requires more surgery.

[skip azp] [skip actions] [skip cirrus]