TST: Fix test_numeric on riscv64
Modify test_numeric so that it passes on riscv64.  The subtest
TestBoolCmp::test_float currently fails on riscv64 as it assumes that
the sign of -np.nan is retained when stored to and read back from an
array.  This is not always the case on riscv64.  Many RISC-V
instructions that produce NaNs return a canonical NaN, as defined by
the RISC-V specification. The canonical NaNs are always positive.  In
this particular test the negative sign of the -np.nan is lost when it
is converted from a double to a float before being stored in
self.signf.  We disable the float32 sign tests for -np.nan on
riscv64 allowing test_numeric to pass.