CI: show meson-log.txt in Cirrus wheel builds [wheel build]
To do this, we need to be able to locate the log file in a fixed
location, hence the need to add `build-dir=build` to the flags
passed to meson-python. And to make that work, it turns out
pypa/build is the better build front-end to use. It's also
going to be the future default for cibuildwheel, so this change
makes sense anyway.

[skip azp] [skip circle]