BLD: optimize BLAS and LAPACK search order
Avoids searching for libraries that cannot be installed for a given
OS or CPU architecture. The introduction of 'auto' as the default
also makes it easier to tweak this later. We reserve the right
to change the behavior of 'auto' as we learn more or as support
for new libraries gets added.