DOC: Improve description of `axis` parameter for `np.median` (#25228)
* DOC: clarify sequence of axes use in numpy.median

The functionality of passing a sequence of axes to the numpy.median axis
parameter is not explained in sufficient detail, leading to possible
confusion about expected results. Specifically, the current description
for the axis parameter does not mention that an array will be flattened
along the given sequence of axes before a median is computed. This
commit changes the description to clarify that the array is first
flattened along the given axes and then a median is computed.
Additionally, a ..versionadded command replaces a written version
acknowledgement.

Also, the docstring for numpy.median does not contain an example for passing a
sequence of ints to the axis parameter. Including an example could help
clarify the functionality of the extended axis feature. This channge
includes a new example demonstrating the median with extended axis.