MAINT: optimization for .replace() method for strings.
The `.replace()` methods counts the number of matches in order
to know the new size. This PR passes on the resulting counts
array to the replace ufunc, since if count is 0, the fast path
can be taken.  Small example of time gain:
```
In [1]: a = np.array(['ABCDEFGHIJKLMNOPQRSTUVWXYZ' * 100] * 10).view(np.char.chararray)

In [2]: %timeit a.replace('0123456', '')
33.2 µs ± 102 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
45.1 µs ± 16.9 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```
As a nice side effect, the string length calculations are now done with
the actual number of replacements that will be done (rather than with
the maximum number, which could lead to overly long strings).