API,MAINT: Reorganize array-wrap calling and introduce ``return_scalar`` (#25409)
This reorganize how array-wrap is called. It might very mildly change the semantics for reductions I think (and for negative priorities).

Overall, it now passes a new return_scalar=False/True when calling __array_wrap__ and deprecates any array-wrap which does not accept arr, context, return_scalar.

I have not integrated it yet, but half the reason for the reorganization is to integrate it/reuse it in the array_coverter helper PR (gh-24214), which stumbled over trying to make the scalar handling sane.

Forcing downstream to add return_scalar=False to the signature is a bit annoying, but e.g. our memory maps currently try to guess at it, which seems bad. I am hoping, this can be part to making the scalar vs. array return more sane.

But, maybe mainly, I hope it consolidates things (together with gh-24124 mainly, as if ufuncs were the only complex place we used this, it wouldn't matter much).

---

* API: Reorganize `__array_wrap__` and add `return_scalar=False`

This also deprecates any `__array_wrap__` which does not accept
`context` and `return_scalar`.

* BUG: Fix niche bug in rounding.

* MAINT: Adjust __array_wrap__ in code and tests (also deprecation test)

* MAINT: Use/move the simplest C-wrapping also

* DOC: Update doc and add release note

* STY: Make linter happy in old tests

* MAINT: Silence GCC warning (value cannot really be used uninitialized)

* MAINT: Small style fixes

* BUG: Fix reference leak in ufunc array-wrapping

This probably doesn't fix the 32bit issue unfortunately, only the windows ones...

* BUG: Fix leak for result arrays in all ufunc calls

* Ensure we try passing context and address related smaller review comments

* Ensure we don't try `context=None` and expand code comment

* Rely on return_scalar always being right (and style nit)

* Remove outdated comments as per review

* Let's just undo force-wrap for now for reductions (its a change...)

* ENH: Chain the original error when the deprecationwarning is raised

Doing this due to gh-25635 since it is super confusing with the
bad retrying...

* BUG,MAINT: Address Martens review comments