ENH: change list-of-array to tuple-of-array returns (Numba compat) (#25570)
Functions in NumPy that return lists of arrays are problematic for
Numba. See https://github.com/numba/numba/issues/8008 for a detailed
discussion on that. This changes the return types to tuples, which
are easier to support for Numba, because tuples are immutable.

This change is not backwards-compatible. Estimated impact:

- idiomatic end user code should continue to work unchanged,
- code that attempts to append or otherwise mutate the list will start
  raising an exception, which should be easy to fix,
- user code that does `if isinstance(..., list):` on the return value of
  a function like `atleast1d` will break. This should be rare, but since
  it may not result in a clean error it is probably the place with the
  highest impact.
- user code with explicit `list[NDArray]` type annotations will need to
  be updated.

Includes some small docstring improvements to make clearer what
is returned.