BLD: fix potential issue with escape sequences in ``__config__.py``
Backport of #25669.

The current code isn't robust to backslashes (from path segments on
Windows) in linker flags. No bug report for NumPy, but the same problem
did show up in SciPy already. Fix is the same as scipy/scipy#19945.

This actually pulls in all of `__config__.py` from main because there
have been a number of other changes.

[skip cirrus] [skip circle]