NEP:  Make a note of Python integer "exceptions"
These are not really exceptions to the rules, but may look like
exceptions when compared to the behavior of other operations.

Also briefly notes that we should probably aspire to allowing such
"exceptions", but I don't think we should be beholden to make it
possible in every case (some of these are non-trivial to make work!).

Also added one sentence that this applies to everything of course.