ENH: add a pkg-config file and a `numpy-config` script
This will make it much easier to support getting at the location
for headers and to support using numpy as a dependency in downstream
build config files. It should enable removing the complexity of
importing numpy and parsing the output of calling `np.get_include` and
co manually. Instead, things should work transparently with (in the
Meson case): `dependency('numpy')`. CMake is also able to consume
pkg-config files and/or a config tool.

Note that the `numpy.pc` file may or may not be installable in the
standard location expected by pkg-config. By default in a wheel it
isn't; distros are in principle able to target the correct location
(however, no flag to do so is added here, it can be done by hand only
right now).

The advantage of the `numpy-config` tool is that it goes into the
scripts directory (i.e., `site-packages/bin`) and hence is always
available. The advantage of the pkg-config file is that it works without
running Python, and hence is better for cross-compiling.

Support for getting at the numpy/f2py/src directory isn't included here,
because that isn't a regular include directory but rather includes
a source file. And hence a compile arg `-I/path/to/numpy/f2py/src`
is not enough. How to better support that requires some more thought.
In its current form, this should already help most packages using the
NumPy C API.