DOC: in ufuncs `dtype` is not ignored when `out` is passed
The documentation stated that the "dtype" argument was not playing any
role when an "out" argument is given. However, that's incorrect, and
both arguments are used. If dtype has a smaller range than the dtype of
out, it can affect the result.

Example:
>>> outi16 = np.empty((1,), numpy.uint16)
>>> np.add.reduce([[200, 200]], axis=1, out=outi16)
array([400], dtype=uint16)
>>> np.add.reduce([[200, 200]], axis=1, out=outi16, dtype=numpy.uint8)
array([144], dtype=uint16)

Also update documentation about multi-output ufuncs.