REV: revert tuple/list return type changes for `*split` functions
This reverts a part of gh-25570. It turns out that the `split`
functions are exceptional, and returning a tuple of arrays isn't
better (perhaps worse) for Numba. For JAX it didn't matter.
Given there's certainly no gain, we decided in the post-merge discussion
on gh-25570 to revert the changes to the `split` APIs.