DOC: Update the CPU build options document
ref: gh-25786

Updated example commands to use ``pip`` instead of ``setup.py``.
Meson doesn't seem to have a command to specify CPU options for ``build_ext`` and ``build_clib`` individually.

The build log (in the ``Build Report`` section) of this document also needs to be updated (gh-24865), but this has not been done yet.

[skip cirrus] [skip azp] [skip actions]