API: Fix compat header and add new import helpers
The current import functions are a bit silly as they are all macros
which immediately return on error.  Thus we really should have a new
helper anyway.

The one added also checks for whether the API was already imported
so that downstream can add it judiciously in case the proper setup
is more difficult.