BUG: fix reinitialization of StringDType after it was set to MISSING.
The current code is meant to track when a string in the arena is freed
for possible re-use later, but actually zeros the offset when an entry
is set to undefined, making this impossible. Furthermore, it does not
track when the offset is reused for either a short string or a string
on the heap, after which the arena should not longer be an option.
This commit fixes both problems, aiming to keep the changes minimally
intrusive.