BUG: Fix ``PyArray_FILLWBYTE`` Cython declaration
The cython declaration was never compatible with the "non deprecated"
API, which didn't matter before because that was incompatible with
NumPy anyway (with Cython 2).
Now it matters, and we want to enforce the new API in this case.

Cython may insert `None` and a type check here for some uses.
I would say this is fine, as it is lightweight and this is only this
one function anyway.

Closes gh-25878