MAINT: remove the experimental `numpy.array_api` module for 2.0
This submodule was added in NumPy 1.22 under an experimental label.
After gaining experience with it, we decided that it was quite
useful for testing code in downstream libraries and from end users
written against the array API standard, but not for production code.
Now that we have array API standard support in the main namespace,
we therefore decided to remove this submodule - as documented in
NEP 56

The code from `numpy.array_api` was moved to the `array-api-strict`
package (see https://github.com/data-apis/array-api-strict) and
further improved there. Users are advised to start using that standalone
package for testing their code if they want to check that it works
not only with `numpy` but with other libraries that support the
array API standard.

This commit also updates the entrypoint (used by for example
`hypothesis`) to point to the main namespace. Users of this entrypoint
to obtain NumPy's compliant namespace should not be affected by this
change.

[skip cirrus]