NEP: update status of active process NEPs, NEP 21 Deferred, NEP 38 Final
We marked process NEPs as Accepted/Final so far, but the intent of
policy/process NEPs that still apply and can be updated when we
update our development processes or project policies is that they
are marked as Active, so they show up under "Meta-NEPs".

NEP 21 is still considered perhaps a good idea, but hasn't been
touched in many years (it's from 2015), so mark it as Deferred.

NEP 38 was finished several years ago, so mark it as Final.

[skip actions] [skip azp] [skip cirrus]