TST: skip rather than xfail a few tests to address CI log pollution
When problematic tests get xfailed, they still run. In our regular CI
jobs we run tests in a way that picks up the `pytest.ini` file in the
root of the repo - that is used to silence noise from such tests.
In the wheel build jobs, this silencing doesn't happen and hence
pytest prints warnings it sees at the bottom of the test output.
In the current wheel builds, this is ~2 pages of warnings.

The solution is straightforward: skip these tests instead (there's not
much point continuing to run them anyway).

For the `f2py` tests I also tried using the `capfd` and `capsys`
fixtures, as is done for other tests. However, a problem is that you
must use one or the other, you can't use both. This may make it
impossible to capture everything.

[skip ci]