API: Modify main dtype/descriptor struct for future dtypes
This modifies the main dtype/descriptor struct to:
* Use intp for the elsize (and alignment)
* A uint64 for flag space
* To actually remove c_metadata and fields related to structured
  dtypes.

It thus *breaks ABI* and the unfortunate souls who require access
to `->elsize` or similar fields will have to vendor `npy_2_compat.h`
or do something similar.
(This should not be super many though, most can use PyArray_ITEMSIZE).

The changes also require moving a few other places to be run-time and
fixing allocation of structs.

This commit is not complete: New warnings and errors still need to be
fixed.