MAINT: Introduce NPY_FEATURE_VERSION_STRING and report it in error
This should also allow checking what a module is compiled for since:
```
strings build/numpy/_core/_rational_tests.cpython-310-darwin.so | grep "NumPy C-API"
```
Will report the error *including* the actual NumPy "python side" version
we are compatible with (so that expanding that grep should be easy for
extracting the interesting value.

Only modules which import NumPy will do this, I suspect that is well enough.