API: revert position-only 'start' in 'np.arange' (#25955)
Closes #25743, reverts #23536

Removing the ability to use start as a kwarg in np.arange(start=3, stop=5) breaks too much even for a 2.0 release. I still am a bit confused about the motivation for the change, was it for typing or automatic signature parsing? Note that np.arange(start=3) raises, and np.arange(stop=5) assumes start=0`.

At some point we may wish to revisit this and remove any use of kwargs, but that would require a long deprecation period.

Edit: any use of kwargs _for start, stop, step