CI: don't use `fetch-tags` in wheel build jobs
This failed in combination with tag-based triggers:
- issue: gh-25979
- cause: https://github.com/numpy/numpy/pull/25981#issuecomment-1987653985

More importantly, we don't need to fetch either the tags or the whole
commit history, because:

(a) we don't derive the version number from the last tag, but from
    parsing pyproject.toml
(b) we don't use a commit count, but only yyyymmdd.git-hash

See `numpy/_build_utils/gitversion.py`

Done only in the wheel build job right now, because that's where the
problem is and we're in a hurry to solve it on the 2.0.x branch.
However, this change should be fine everywhere.

[skip cirrus] [skip circle] [skip azp]