MAINT: handle `NPY_ALLOW_THREADS` and related build option better
The option to disable threading in CPython was removed a long time ago
(see cpython#3385); `WITH_THREAD` was kept defined for backwards
compatibility, but can never change - so best to remove checking for
it.

The docs still mentioned the old `NPY_NOSMP` environment variable,
support for which was removed with the move to Meson. Instead, there
is a `disable-threading` build option, so document that.