MAINT: Make PyArrayMultiIterObject struct "smaller"
This inverts the lie, on C++ we claim the struct is only fixed
to 32 iterators, while in reality such a limit doesn't exist.

The reason is that Cython is unhappy if it finds a smaller struct
at runtime (on NumPy 1.x).

If there is more oddities around this, we could probably even detect
Cython...