API: Default to hidden visibility for API tables
This adds a new ``NPY_API_SYMBOL_ATTRIBUTE`` but defaults to using
hidden visibility (always the case on windows!).

That actually makes the situation "worse" for ``eigenpy`` in some degree,
since it forces them to adapt, but it also allows them to decide to just
roll with it (by actually also exporting it on windows which).

Since it aligns windows and linux it seems like a good idea?  OTOH,
who knows how many projects get away with just not caring about windows...

I tried to reorganize the docs a bit on how to import the array API...