BUG,MAINT: Fix __array__ bugs and simplify code
This fixes that the UserWarning is incorrect, it needs to be a
DeprecationWarning (it does not related to end-users).
Instead, include the information about `copy=None` into the error
message.

It also fixes a bug:  If `copy=` kwarg is unsupported we cannot
guarantee no-copy, so we should raise, or did we discuss to "just warn"?

That is also a latent bug: When `copy=True` and the fallback path
is taken, then we must make a copy to be on the safe side.