MNT: replace _PyDict_GetItemStringWithError with PyDict_GetItemStringRef (#26282)
Replaces all usages of the private _PyDict_GetItemStringWithError Python C API function with PyDict_GetItemStringRef. The latter is available in the python 3.13 C API and via python-capicompat in older python versions.

The new function has the same semantics as PyDict_GetItemRef but takes a UTF-8 C string instead of a python object. Like _PyDict_GetItemStringWithError, it does not suppress errors. It also has the nice improvement that the return type is now int and it signals success or failure. This lets me re-structure control flow a bit and avoid PyErr_Occurred() calls.

I used the new PyDict_ContainsString function if all we care about is whether a key is in the dict. I also used Py_SETREF in a few places that were using the unsafe decref then set pattern.

See #26159 for the relevant tracking issue.

---

* MNT: replace _PyDict_GetItemStringWithError with PyDict_GetItemStringRef

* MNT: fix issue spotted in code review

* MNT: fix two more issues spotted in code review

* MNT: more missing decrefs spotted in code review

* MNT: delete unnecessary decrefs (covered by goto fail logic)