MAINT: improper error handling and cleanup.
I just came across that while trying to benchmark docs.

It seem that when this was moved from subprocess.run to spin.util run the
error handling was not updated.

spin.util seem to always return a CompletedProcess and it has the stderr
and return code. (tested locally by changing towncrier to `false`).

While at it I removed unused import and reformatted a tiny bit the
pyprojet.toml to only have 1 command per line for redability.